Replace float by mreal everywhere (if possible) for higher accuracy of time ticks

+++ src/def_font.cpp
@@ -19,7 +19,7 @@
  ***************************************************************************/
 //# font STIXGeneral.otf has 516 glyphs
 unsigned mgl_numg=516, mgl_cur=246080;
-float mgl_fact=50;
+double mgl_fact=50;
 long mgl_gen_fnt[516][6] = {
 	{33, 232, 41, 0, 39, 62042},
 	{34, 285, 41, 82, 38, 62276},
